requirements.txt
llama-index
google-generativeai
sentence-transformers
python-dotenv
fastapi
uvicorn
python-multipart
faiss-cpu
llama-index-llms-gemini